feat(preview ui): add details card
