+++ package.json
@@ -43,6 +43,7 @@
     "crypto-js": "^4.1.1",
     "date-fns": "2.28.0",
     "lodash": "^4.17.21",
+    "moment": "^2.30.1",
     "next": "^13.3.0",
     "next-auth": "^4.24.7",
     "next-connect": "^1.0.0-next.4",
