+++ src/utils/date-utils.tsx
@@ -1,0 +1,5 @@
+import moment from 'moment';
+
+export const getLastNthDate = (n: number) => {
+  return moment().subtract(n, 'days').format('DD-MM-YYYY');
+};
