+++ src/utils/explore-utils.tsx
@@ -12,3 +12,18 @@ export const generateExplorePayload = (wid: string, pid: string, user: any) => {
     }
   };
 };
+
+const promptFilters = [
+  { val: 7, name: 'Last 7 days' },
+  { val: 15, name: 'Last 15 days' },
+  { val: 30, name: 'Last 30 days' },
+  { val: 90, name: 'Last 90 days' },
+  { val: 60, name: 'Last 2 months' }
+];
+
+export const getPromptFilter = (filter: string) => {
+  //@ts-ignore
+  const { val } = promptFilters.find((item) => filter === item.name);
+
+  return val;
+};
